+++ Source/Tools/ModelBaker.cpp
@@ -256,7 +256,7 @@ void ModelBaker::BakeFiles(FileCollection files)
         }));
     }
 
-    int32 errors = 0;
+    size_t errors = 0;
 
     for (auto& file_baking : file_bakings) {
         try {
